# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EDH Technology API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts:
    post:
      operationId: submitContact
      tags: [contacts]
      summary: Submit a contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contact submissions
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - subject
        - message
        - createdAt
    ContactInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - name
        - email
        - subject
        - message
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error